requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
